==> combined_data_without_soh/analysis/03b_Phylogenetic_Tree_meta_analysis.Rmd <==
---
title: "Phylogenetic Tree Inspection and Rooting - meta analysis"
author: "Calla Bush St George"
date: "`r Sys.Date()`"
output:
  html_document: 
    code_folding: show
    theme: spacelab
    highlight: pygments
    keep_md: no
    toc: yes
    toc_float:
      collapsed: no
      smooth_scroll: yes
      toc_depth: 3
  keep_md: true  
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      fig.align = "center",
                      fig.path = "../figures/03_Phylogenetic_Tree/")
```

# Goals

1. Load the FastTree unrooted tree.
2. Add tree to phyloseq object.
3. Visualize and inspect tree with ggtree.
4. Prune ASVs, if needed.
5. Root our tree.
6. Combine new tree with a phyloseq objet.
7. Save 2 phyloseq objects: 1. Unrooted tree phyloseq object, 
                            2. Rooted tree phyloseq object.

# Setting the Environment

## Set the seed
```{r set-seed}
set.seed(238428)
```

## Load packages
```{r load-packages}
pacman::p_load(tidyverse, phyloseq, ggtree, phytools, 
               install = FALSE)
```

## Load Data files
```{r load-data}
# Preprocessed phyloseq object
load("combined_data/data/02_PreProcessing/raw_preprocessed_physeq.RData")
raw_preprocessed_physeq

# Load in the tree!
unrooted_tree <- 
  read.tree("combined_data/data/03_Phylogenetic_Tree/ASVs_unrooted.tree")
unrooted_tree
str(unrooted_tree)
```

# Load colors
```{r phylum-colors}
# Set the phylum colors
phylum_colors <- c(
  "navy", 
  "darkslategray2", 
  "deeppink1",
  "plum2", 
  "gold", 
  "red1",
  "black", 
  "firebrick",
  "dodgerblue",
  "limegreen",
  "aquamarine",
  "mediumpurple4",
  "magenta4",
  "grey", 
  "magenta",
  "#3E9B96",
  "pink4",
  "greenyellow",
  "red4",
  "yellow",
  "#B5D6AA",
  "palevioletred1",
  "royalblue",
  "darkorange", 
  "olivedrab",
  "indianred",
  "blue1",
  "greenyellow",
  "orange4",
  "green",
  "darkorchid1",
  "orange",
  "blue4",
  "sienna",
  "seagreen",
  "yellow4",
  "thistle",
  "lightgreen")
length(phylum_colors)
```


# Merge Phyloseq
```{r merge-physeq}
ntaxa(raw_preprocessed_physeq)
ntaxa(unrooted_tree)

# Merge the tree with the phyloseq object
unrooted_physeq <-
  merge_phyloseq(raw_preprocessed_physeq, unrooted_tree)

unrooted_physeq
```

```{r plot-tree-unrooted}
# Make a basic tree
kingdom_node_tree <-
  ggtree(unrooted_physeq) +
  # Color by kingdom
  geom_tippoint(mapping = aes(color = Kingdom)) +
  scale_color_manual(values = c("goldenrod1", "cornflowerblue", "gray")) +
  # Add title
  labs(title = "Unrooted Tree") + 
  # Move legened to the bottom
  theme(legend.position = "bottom") +
  # Add node labels
  geom_text(aes(label = ASV), hjust = 0.5, vjust = 4, size = 3)
kingdom_node_tree
```


# Prune ASV_3354
```{r prune-ASV-3354}
# Function from Joey McMurdie: https://github.com/joey711/phyloseq/issues/652
pop_taxa = function(physeq, badTaxa){
  allTaxa <-  taxa_names(physeq)
  allTaxa <- allTaxa[!(allTaxa %in% badTaxa)]
  return(prune_taxa(allTaxa, physeq))
  }

# Let's use the pop taxa function :)
# Recreate a phyloseq object without ASV_456
unrooted_physeq_rm3354 <-
  unrooted_physeq %>%
  pop_taxa(., "ASV_3354")

# Intuition Check
ntaxa(unrooted_physeq) - ntaxa(unrooted_physeq_rm3354)

# Visually inspect
ggtree(unrooted_physeq_rm3354) +
  geom_tippoint(mapping = aes(color = Kingdom))+
  scale_color_manual(values = c("goldenrod1", "cornflowerblue", "gray"))
```


# Midroot Tree

```{r midroot-tree}
# Is the tree rooted? 
new_unrooted_tree <- phy_tree(unrooted_physeq_rm3354)

# Let's midpoint root the tree
midpoint_rooted_tree <-
  midpoint.root(new_unrooted_tree)

# Assign to a new phyloseq object: merging subsetted phyloseq with the new rooted tree

# 1. Create the phyloseq object
physeq_rm3354 <-
  raw_preprocessed_physeq %>%
  subset_taxa(ASV != "ASV_3354")

# Merge tree with the new physeq_rm456
midroot_physeq <-
  merge_phyloseq(physeq_rm3354, midpoint_rooted_tree)
midroot_physeq

# Quick Inspection of tree
ggtree(midroot_physeq) + 
  geom_tippoint(mapping = aes(color = Kingdom))+
  scale_color_manual(values = c("goldenrod1", "cornflowerblue", "gray"))

ggtree(midroot_physeq) + 
  geom_tippoint(mapping = aes(color = Phylum))+
  scale_color_manual(values = phylum_colors)
```

# Save to a new phyloseq object

```{r save-physeq}
# Save both phyloseq objects with our tree object to one .RData file
save(list = c("unrooted_physeq", "midroot_physeq"),
     file = "combined_data/data/03_Phylogenetic_Tree/phytree_preprocessed_physeq.RData")
```



# Session Information
```{r session-info}
devtools::session_info()
```
